Move source utils into `features/sources/`
